pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Transcription of calls from trunk-recorder using OpenAI Whisper"
authors = ["Eric Tendian"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.26"
boto3-stubs = "^1.26"
celery = "^5.3"
fastapi = "^0"
flower = "^2.0"
meilisearch = "^0"
python-dotenv = "^0"
python-multipart = "^0"
pytz = "*"
requests = "^2.28"
uvicorn = {extras = ["standard"], version = "^0"}
apprise = "^1.2"
cachetools = "^5.3"
sentry-sdk = {extras = ["celery", "fastapi"], version = "^1.14"}
openai = "^1.24"
python-datauri = "^1.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
sqlalchemy = "^2.0.22"
geopy = "^2.4.1"
google-generativeai = "^0.4.0"
deepgram-sdk = "^3.2.2"
shapely = "^2.0.4"
requests-cache = "^1.2.0"
celery-batches = "^0.8.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.2.5"
pytest = "^8.3.2"
mypy = "^1.11.2"
ruff = "^0.6.3"
celery-stubs = "^0.1.3"
types-requests = "^2.32.0.20240712"
types-cachetools = "^5.5.0.20240820"
types-pytz = "^2024.1.0.20240417"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = false
exclude = ["bin", "whisper"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "E402",  # Module level import not at top of file
]

[tool.ruff.lint.per-file-ignores]
"app/bin/scratch/*" = ["E"]
